content-management-service/requirements.txt
fastapi==0.115.12
uvicorn
SQLAlchemy==2.0.34
demjson3==3.0.6
requests==2.32.3
python-dotenv==1.0.1
Werkzeug==3.0.3
pg8000==1.31.2
psycopg2-binary==2.9.9
celery
tiktoken
openai
textract==1.6.3
python-multipart